+++ admin/app/api/keystatic/[...params]/route.ts
@@ -1,8 +1,0 @@
-import path from 'node:path';
-import { makeRouteHandler } from '@keystatic/next/route-handler';
-import config from '../../../../keystatic.config';
-
-export const { POST, GET } = makeRouteHandler({
-  config,
-  localBaseDirectory: path.resolve(process.cwd(), '..'),
-});